refactor: extract search queries from resumes.ts into resumes_search.ts
Extract all search-related query/mutation/action definitions from
resumes.ts (2022 → 1171 lines, 42% reduction) into a new focused
resumes_search.ts module (926 lines).

Moved exports: search, searchWithIngestData, searchWithTagExpansion,
searchWithTagExpansionPage, searchWithTagExpansionPaginated,
searchWithTagExpansionScanPage, searchWithTagExpansionAndMode,
scanResumePageByTime, scanResumePageSlim, getResumes,
getResumeDocsByIds, collectSearchIndexDocIds, getResumesByIds.

Also moved private helpers: compareResumes, mergeResumeDocs,
resolveSearchWithTagExpansionMatch, runSearchWithTagExpansionPageQuery,
runSearchWithTagExpansionScanPageQuery.

Updated all callers: BFF string-path references (resumes: →
resumes_search:), frontend api.resumes.* → api.resumes_search.*,
internal Convex references (internal.resumes.getResumesByIds →
internal.resumes_search.getResumesByIds), and test files.
